Fix NMEA GSV decoding
On a GNSS with a disconnected anntenna, the GNSS managed to report
seeing a single satellite. However, it failed to report the elevation
and azimuth. It's not clear from the NMEA specification if this dataset
is valid, but we should handle the situation. While here, return the
elevation, azimuth, and SNR as strings. These values are decimal64 and
a string is the correct encoding for RFC7951.

Bugfix: VRVDR-55360