Check TDPLL input status for passive antenna
If we have a splitter to split antenna signal and feed to 2 SIADs,
we might can only see a passive antenna from one of SIADs.
We cannot distinguish Open and passive antenna,  so only read antenna
status from CPLD is not enough.  We also need to check TDPLL input
status.  If we have a valid GPS input from TDPLL side, we treat it as
antenna 'OK'.

VRVDR-55418