UbloxGNSS: handle missing fields
If the position or time isn't known, then those sentences will have blank
fields. Previously, the try/except caught these errors but this resulted
in early exits during parsing. While the time might not be valid yet,
the list of known satellites could still be valid.

Restrict the try/catch to just handling USB errors and manually handle
the missing time and position data. Also, ensure that we get a full set
of records for the satellite data.

Bugfix: VRVDR-55804