UBloxGNSS: perform a hardware reset on stop
Stopping the GNSS tasks does not cause the 10Mhz signal to stop.
The presence of that signal prevents some platforms from switching
to a backup timing source. To fix this problem, perform a hardware
reset. Note that this is a hard reset, so the hardware will be briefly
unavailable. After the hardware returns, we then issue the controlled
stop to keep the GNSS idle. Note that this will stop and start will
force the GNSS to completely reacquire a satellite lock from scratch.

Bugfix: VRVDR-55589