ublox: use UBX-NAV-SAT to report satellites
The stream of GSV records appear to compete with the other messages
resulting in lost command acknowledgements. Switch to the UBX-NAV-SAT
message to get the satellites in view. While here, stop using RMC records
and use UBX-NAV-PVT to get the position and time. This change eliminates
the polling reads of the GNSS receiver.

Bugfix: VRVDR-56560